+++ resources/aws_connect_approved_origin.rb
@@ -1,0 +1,42 @@
+# Import API specifics
+use "awscc_base"
+
+resource_name :aws_connect_approved_origin
+provides :aws_connect_approved_origin, target_mode: true, platform: "aws"
+
+description <<~DESCRIPTION
+  Resource Type definition for AWS::Connect::ApprovedOrigin
+DESCRIPTION
+
+property :name, String,
+         name_property: true,
+         description: "Name of the resource, not desired state"
+
+property :instance_id, Hash,
+         required: true,
+         callbacks: {
+           "instance_id is not a String" => lambda { |v| v.is_a? String },
+           "instance_id needs to be 1..100 characters" => lambda { |v| v.length >= 1 && v.length <= 100 },
+           "instance_id must match pattern ^arn:aws[-a-z0-9]*:connect:[-a-z0-9]*:[0-9]{12}:instance/[-a-zA-Z0-9]*$" => lambda { |v| v =~ Regexp.new("/^arn:aws[-a-z0-9]*:connect:[-a-z0-9]*:[0-9]{12}:instance/[-a-zA-Z0-9]*$/") },
+         },
+         description: ""
+
+property :origin, Hash,
+         required: true,
+         callbacks: {
+           "origin is not a String" => lambda { |v| v.is_a? String },
+         },
+         description: ""
+
+# API URLs and mappings
+rest_api_collection "/AWS::Connect::ApprovedOrigin"
+rest_api_document "/AWS::Connect::ApprovedOrigin"
+
+rest_property_map({
+  instance_id: "InstanceId",
+  origin:      "Origin",
+})
+
+rest_post_only_properties %i{
+  instance_id origin
+}